catch NullPointerException where symbolcode is missing, change to NotNullMap in getSymbolCode-methods LF repo

+++ project/app/src/main/java/com/example/myapplication/ui/surfarea/SurfAreaScreen.kt
@@ -96,8 +96,8 @@ fun SurfAreaScreen(
                         val surfAreaDataForDay = nextSevenDays.getOrElse(dayIndex) { emptyList() }
                         var maxWaveHeight = 0.0
                         surfAreaDataForDay.forEach { surfAreaDataForHour ->
-                            if (maxWaveHeight < surfAreaDataForHour.second[0]) {
-                                maxWaveHeight = surfAreaDataForHour.second[0]
+                            if (maxWaveHeight < surfAreaDataForHour.second[0] as Double) {
+                                maxWaveHeight = surfAreaDataForHour.second[0] as Double
                             }
                         }
